size should be int

+++ src/main.cpp
@@ -20,7 +20,7 @@ static void removeFiles(const std::string& path) {
     }
 }
 
-static void downloadTitlesFromArray(Json::Value array, size_t size) {
+static void downloadTitlesFromArray(Json::Value array, int size) {
     for (int i = 0; i < size; ++i) {
         char *outputPath = strdup(array[i].get("path", "./mlc01/sys/title/00050010/1004A200").asCString());
         downloadTitle(array[i].get("titleID", "000500101004A200").asCString(), outputPath);
